+++ serde_with/src/duplicate_key_impls/first_value_wins.rs
@@ -1,63 +1,16 @@
-use alloc::collections::{BTreeMap, BTreeSet};
+use alloc::collections::BTreeMap;
 #[cfg(any(feature = "std", feature = "indexmap"))]
 use core::hash::{BuildHasher, Hash};
 #[cfg(feature = "indexmap")]
 use indexmap::IndexMap;
 #[cfg(feature = "std")]
-use std::collections::{HashMap, HashSet};
-
-#[deprecated = "This is serde's default behavior."]
-pub trait DuplicateInsertsFirstWinsSet<T> {
-    fn new(size_hint: Option<usize>) -> Self;
-
-    /// Insert the value into the set, if there is not already an existing value
-    fn insert(&mut self, value: T);
-}
+use std::collections::HashMap;
 
 pub trait DuplicateInsertsFirstWinsMap<K, V> {
     fn new(size_hint: Option<usize>) -> Self;
 
     /// Insert the value into the map, if there is not already an existing value
     fn insert(&mut self, key: K, value: V);
-}
-
-#[cfg(feature = "std")]
-#[allow(deprecated)]
-impl<T, S> DuplicateInsertsFirstWinsSet<T> for HashSet<T, S>
-where
-    T: Eq + Hash,
-    S: BuildHasher + Default,
-{
-    #[inline]
-    fn new(size_hint: Option<usize>) -> Self {
-        match size_hint {
-            Some(size) => Self::with_capacity_and_hasher(size, S::default()),
-            None => Self::with_hasher(S::default()),
-        }
-    }
-
-    #[inline]
-    fn insert(&mut self, value: T) {
-        // Hashset already fulfils the contract and always keeps the first value
-        self.insert(value);
-    }
-}
-
-#[allow(deprecated)]
-impl<T> DuplicateInsertsFirstWinsSet<T> for BTreeSet<T>
-where
-    T: Ord,
-{
-    #[inline]
-    fn new(_size_hint: Option<usize>) -> Self {
-        Self::new()
-    }
-
-    #[inline]
-    fn insert(&mut self, value: T) {
-        // BTreeSet already fulfils the contract and always keeps the first value
-        self.insert(value);
-    }
 }
 
 #[cfg(feature = "std")]
