+++ serde_with/src/duplicate_key_impls/mod.rs
@@ -2,9 +2,8 @@ mod error_on_duplicate;
 mod first_value_wins;
 mod last_value_wins;
 
-#[allow(deprecated)]
 pub use self::{
     error_on_duplicate::{PreventDuplicateInsertsMap, PreventDuplicateInsertsSet},
-    first_value_wins::{DuplicateInsertsFirstWinsMap, DuplicateInsertsFirstWinsSet},
+    first_value_wins::DuplicateInsertsFirstWinsMap,
     last_value_wins::DuplicateInsertsLastWinsSet,
 };
