+++ serde_with/src/guide.md
@@ -8,7 +8,7 @@ For example, you can serialize [a map as a sequence of tuples][crate::guide::ser
 
 The crate offers four types of functionality.
 
-## 1. A more flexible and composable replacement for the with annotation, called `serde_as` *(v1.5.0+)*
+## 1. A more flexible and composable replacement for the with annotation, called `serde_as`
 
 This is an alternative to [serde's with-annotation][with-annotation], which adds flexibility and composability to the scheme.
 The main downside is that it work with fewer types than [with-annotations][with-annotation].
@@ -62,63 +62,7 @@ assert_eq!(json, serde_json::to_string_pretty(&data).unwrap());
 assert_eq!(data, serde_json::from_str(json).unwrap());
 ```
 
-## 2. Integration with serde's with-annotation
-
-[serde's with-annotation][with-annotation] allows specifying a different serialization or deserialization function for a field.
-It is useful to adapt the serialization of existing types to the requirements of a protocol.
-Most modules in this crate can be used together with the with-annotation.
-
-The annotation approach has one big drawback, in that it is very inflexible.
-It allows specifying arbitrary serialization code, but the code has to perform the correct transformations.
-It is not possible to combine multiple of those functions.
-One common use case for this is the serialization of collections like `Vec`.
-If you have a field of type `T`, you can apply the with-annotation, but if you have a field of type `Vec<T>`, there is no way to re-use the same functions for the with-annotation.
-This inflexibility is fixed in the `serde_as` scheme presented above.
-
-The example shows a similar setup as in the `serde_as` example above, but using the with-annotation.
-
-### Example
-
-```rust
-# use serde::{Deserialize, Serialize};
-# use std::net::Ipv4Addr;
-#
-# #[derive(Debug, PartialEq, Eq)]
-#[derive(Deserialize, Serialize)]
-struct Data {
-    // Type does not implement Serialize or Deserialize
-    #[serde(with = "serde_with::rust::display_fromstr")]
-    address: Ipv4Addr,
-    // Treat the Vec like a map with duplicates
-    #[serde(with = "serde_with::rust::tuple_list_as_map")]
-    vec_as_map: Vec<(String, u32)>,
-}
-
-let data = Data {
-    address: Ipv4Addr::new(192, 168, 0, 1),
-    vec_as_map: vec![
-        ("Hello".into(), 123),
-        ("World".into(), 456),
-        ("Hello".into(), 123),
-    ],
-};
-
-let json = r#"{
-  "address": "192.168.0.1",
-  "vec_as_map": {
-    "Hello": 123,
-    "World": 456,
-    "Hello": 123
-  }
-}"#;
-
-// Test Serialization
-assert_eq!(json, serde_json::to_string_pretty(&data).unwrap());
-// Test Deserialization
-assert_eq!(data, serde_json::from_str(json).unwrap());
-```
-
-## 3. proc-macros to make it easier to use both above parts
+## 2. proc-macros to make it easier to use both above parts
 
 The proc-macros are an optional addition and improve the user experience for common tasks.
 We have already seen how the `serde_as` attribute is used to define the serialization instructions.
@@ -127,15 +71,10 @@ The proc-macro attributes are defined in the [`serde_with_macros`] crate and re-
 The proc-macros are optional, but enabled by default.
 For further details, please refer to the documentation of each proc-macro.
 
-## 4. Derive macros to implement `Deserialize` and `Serialize`
+## 3. Derive macros to implement `Deserialize` and `Serialize`
 
 The derive macros work similar to the serde provided ones, but they do implement other de/serialization schemes.
 For example, the derives [`DeserializeFromStr`] and [`SerializeDisplay`] require that the type also implement [`FromStr`] and [`Display`] and de/serializes from/to a string instead of the usual way of iterating over all fields.
-
-## Migrating from the with-annotations to `serde_as`
-
-Each old style module explains how it can be converted to `serde_as`.
-Not all modules have such a description since not all are migrated and some are hard to implement in the `serde_as` system.
 
 [`Display`]: std::fmt::Display
 [`FromStr`]: std::str::FromStr
